+++ includes/HookHandlers/Main.php
@@ -5,18 +5,24 @@ namespace Miraheze\RottenLinks\HookHandlers;
 use JobQueueGroup;
 use MediaWiki\Deferred\LinksUpdate\LinksUpdate;
 use MediaWiki\Hook\LinksUpdateCompleteHook;
+use MediaWiki\Hook\ParserFirstCallInitHook;
+use MediaWiki\Parser\Parser;
 use Miraheze\RottenLinks\RottenLinksJob;
+use Miraheze\RottenLinks\RottenLinksParserFunctions;
+use Wikimedia\Rdbms\IConnectionProvider;
 
-class Main implements LinksUpdateCompleteHook {
+class Main implements LinksUpdateCompleteHook, ParserFirstCallInitHook {
 
-	/** @var JobQueueGroup */
-	private $jobQueueGroup;
+	private JobQueueGroup $jobQueueGroup;
+	private RottenLinksParserFunctions $parserFunctions;
 
 	/**
 	 * @param JobQueueGroup $jobQueueGroup
+	 * @param IConnectionProvider $connectionProvider
 	 */
-	public function __construct( JobQueueGroup $jobQueueGroup ) {
+	public function __construct( JobQueueGroup $jobQueueGroup, IConnectionProvider $connectionProvider ) {
 		$this->jobQueueGroup = $jobQueueGroup;
+		$this->parserFunctions = new RottenLinksParserFunctions( $connectionProvider );
 	}
 
 	/**
@@ -37,5 +43,14 @@ class Main implements LinksUpdateCompleteHook {
 
 			$this->jobQueueGroup->push( new RottenLinksJob( $params ) );
 		}
+	}
+
+	/**
+	 * Handler for ParserFirstCallInit hook.
+	 * @see https://www.mediawiki.org/wiki/Manual:Hooks/ParserFirstCallInit
+	 * @param Parser $parser
+	 */
+	public function onParserFirstCallInit( $parser ) {
+		$parser->setFunctionHook( 'rl_status', [ $this->parserFunctions, 'onRLStatus' ], Parser::SFH_OBJECT_ARGS );
 	}
 }