+++ includes/mw.ext.rottenLinks.lua
@@ -1,0 +1,18 @@
+local p = {}
+
+function p.setupInterface(arguments)
+	p.setupInterface = nil
+
+	local php = mw_interface
+	mw_interface = nil
+
+	mw = mw or {}
+	mw.ext = mw.ext or {}
+
+	p.getStatus = php.getStatus
+
+	mw.ext.rottenLinks = p
+	package.loaded['mw.ext.rottenLinks'] = p
+end
+
+return p